requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
msgpack==1.1.0
pillow==11.2.1
python-decouple==3.8
redis==6.0.0
sqlparse==0.5.3
tzdata==2025.2
